+++ RELEASE_NOTES.md
@@ -1,8 +1,15 @@
 Release Notes
 =============
+## 1.0.0
+* Formal release
+
+## 1.0.0 RC2
+* Fix an issue with CosmosDB tags being set twice
+* ACI dropped support for assigning static private IP
+
 ## 1.0.0 RC
 * Postgres API redesign
-* Network Security Group API design
+* Network Security Group API redesign
 * Validation for all Storage Account resources
 * Prevent supplying VM custom script files without a custom script
 * Basic Validation helper functions